refactor : 논의방 infra계층 폴더 구조변경
- /member 경로 생성
- /discussionRoom 경로 생성
- 기존 6개의 파일 각 3개씩 위 둘 경로로 이동
